Make sse prefixes opcode extensions
This change makes sse prefixes (that disambiguate 2+ opcode bytes
instructions) act like rest of the opcode extensions. Importantly,
see prefix tables are now created towards the lower part of the
opcode tree.

However, note that SSE prefixes *must* be evaluated before the
operand size based disambiguation. This is because one the sse
prefix is consumed, it is discarded. So, in the case of the 66
prefix, which doubles as operand size override, it must be
consumed before operand size disambiguation.