Fixes for assembly output buffer
As pointed out by Andrew Sveikauskas (https://github.com/asveikau),
mkasm does not check for overflows. This change address the overflow
issue in a simple way, by dropping/truncating the output and increasing
the buffer size to 128 chars to accomodate most common cases. If a
client anticipates larger output requirements, the change adds a new
interface ud_set_asm_buffer, that can switch to a user buffer. The
behavior is still the same, output is truncated on overflow.