Fix string prefix decoding
Issue #10 points out a flaw in decoding the string prefixes in
udis86. Although rep/repe use the same encoding, the precise
prefix depends on the isntruction.

This change adds a new prefix notation to optable.xml that
indicates if an instruction takes a rep or repz (rep with
zero-flag).

It is undefined in the isa what the behavior of this 0xf3
prefix would on instructions that do not support it. For now,
udis86 simply decodes it as rep.