Simplify standalone header include contract
(Issue #62) In __UD_STANDALONE__ mode, the onus is now on the client to
make sure that the right set of files are included prior to including
udis86.h in their environment. types.h no longer assumes the
availability of header files. I haven't tried this myself in a kernel
build environment, but I'm hoping someone will, and get back with
problems if any.