No UD_ATTRIBUTE_PACKED for ud_mnemonic_code
I don't think packing the enum into 16-bit is advantageous in any way, also it doesn't happen on MSVC, for which the macro is defined to be empty. The ud struct's size shouldn't change (at least it doesn't on Clang/GCC for 64-bit), only the offset of the mnemonic member would be aligned to 4 bytes.